requirements.txt
# TODO Reduce dependency
PyMuPDF
langchain-community
sentence_transformers
sentencepiece
rank-bm25
scikit-learn 
tiktoken 
torch
umap-learn  
unstructured 
nltk   
transformers 
pymilvus  
tqdm 
openai
streamlit
langchain_huggingface